Moved agent state tests to their own file.
